fullhistory: fix golangci-lint findings
golangci-lint job (default: all) on the orchestration + daemon layer:
- gocognit: extract awaitCoverageBuilds out of executePlan
- nestif: extract confirmEarliestUnchanged out of validateConfig
- minmax: use max() in lastCommittedLedger floor + IngestionLag
- intrange: range-over-int in networkTip's retry loop
- nolintlint: drop three now-unused //nolint:gosec directives
- misspell: cancelled -> canceled
- lll: wrap/shorten the over-long lines
- tests: nonamedreturns, contextcheck, unparam, prealloc, unconvert,
  embeddedstructfieldcheck, protogetter, and testifylint float-compare (InDelta)

+++ cmd/stellar-rpc/internal/fullhistory/helpers_test.go
@@ -26,7 +26,7 @@ import (
 
 // testCPI is the tx-hash index width tests build layouts with; equals the
 // production constant so on-disk geometry reads back identically.
-const testCPI = uint32(geometry.ChunksPerTxhashIndex)
+const testCPI = geometry.ChunksPerTxhashIndex
 
 func silentLogger() *supportlog.Entry {
 	var buf bytes.Buffer
@@ -172,6 +172,7 @@ func (s *fullChunkStream) RawLedgers(
 // LedgerStream (fullChunkStream) plus a programmable frontier Tip.
 type fakeBackend struct {
 	ledgerbackend.LedgerStream
+
 	tip uint32
 }
